(PC-35819) Fix(CI):  handle some sonar issue on Circular file imports
