+++ src/features/artist/pages/Artist.tsx
@@ -2,17 +2,11 @@ import { useRoute } from '@react-navigation/native'
 import React, { FunctionComponent, useMemo } from 'react'
 
 import { ArtistBody } from 'features/artist/components/ArtistBody/ArtistBody'
+import { ArtistType } from 'features/artist/types'
 import { UseRouteType } from 'features/navigation/RootNavigator/types'
 import { useArtistResults } from 'features/offer/helpers/useArtistResults/useArtistResults'
 import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
-
-export type Artist = {
-  id: string
-  name: string
-  bio?: string
-  image?: string
-}
 
 export const Artist: FunctionComponent = () => {
   const enableArtistPage = useFeatureFlag(RemoteStoreFeatureFlags.WIP_ARTIST_PAGE)
@@ -31,7 +25,7 @@ export const Artist: FunctionComponent = () => {
   // TODO(PC-35430): replace null by PageNotFound when wipArtistPage FF deleted
   if (!artist || !enableArtistPage) return null
 
-  const artistInfo: Artist = {
+  const artistInfo: ArtistType = {
     id: params.id,
     name: artist.name,
     image: artist.image,
