+++ src/features/artist/types.ts
@@ -1,0 +1,6 @@
+export type ArtistType = {
+  id: string
+  name: string
+  bio?: string
+  image?: string
+}
