+++ src/features/home/components/modules/business/NewBusinessModule.native.test.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { Linking } from 'react-native'
 
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { BusinessModuleProps } from 'features/home/components/modules/business/BusinessModule'
 import { NewBusinessModule } from 'features/home/components/modules/business/NewBusinessModule'
+import { BusinessModuleProps } from 'features/home/types'
 import { beneficiaryUser } from 'fixtures/user'
 import { analytics } from 'libs/analytics/provider'
 import { ContentTypes } from 'libs/contentful/types'
