+++ src/features/home/components/modules/business/NewBusinessModule.tsx
@@ -4,9 +4,9 @@ import LinearGradient from 'react-native-linear-gradient'
 import styled from 'styled-components/native'
 
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { BusinessModuleProps } from 'features/home/components/modules/business/BusinessModule'
 import { getBusinessUrl } from 'features/home/components/modules/business/helpers/getBusinessUrl'
 import { useShouldDisplayBusinessModule } from 'features/home/components/modules/business/helpers/useShouldDisplayBusinessModule'
+import { BusinessModuleProps } from 'features/home/types'
 import { openUrl } from 'features/navigation/helpers/openUrl'
 import { AccessibilityRole } from 'libs/accessibilityRole/accessibilityRole'
 import { analytics } from 'libs/analytics/provider'
