+++ src/features/home/components/modules/exclusivity/ExclusivityExternalLink.tsx
@@ -2,7 +2,7 @@ import React, { memo, useEffect } from 'react'
 import styled from 'styled-components/native'
 
 import { ExclusivityImage } from 'features/home/components/modules/exclusivity/ExclusivityImage'
-import { ExclusivityBannerProps } from 'features/home/components/modules/exclusivity/ExclusivityModule'
+import { ExclusivityBannerProps } from 'features/home/types'
 import { analytics } from 'libs/analytics/provider'
 import { ContentTypes } from 'libs/contentful/types'
 import { useHandleFocus } from 'libs/hooks/useHandleFocus'
