+++ src/features/home/components/modules/exclusivity/ExclusivityModule.tsx
@@ -5,23 +5,8 @@ import styled from 'styled-components/native'
 import { ExclusivityBanner } from 'features/home/components/modules/exclusivity/ExclusivityBanner'
 import { ExclusivityExternalLink } from 'features/home/components/modules/exclusivity/ExclusivityExternalLink'
 import { ExclusivityOffer } from 'features/home/components/modules/exclusivity/ExclusivityOffer'
-import { ExclusivityModule as ExclusivityModuleType } from 'features/home/types'
+import { ExclusivityModuleProps } from 'features/home/types'
 import { Spacer } from 'ui/theme'
-
-export interface ExclusivityModuleProps {
-  title: string
-  alt: string
-  image: string
-  moduleId: string
-  offerId?: number
-  displayParameters?: ExclusivityModuleType['displayParameters']
-  url?: string
-  homeEntryId: string | undefined
-  index: number
-  style?: ViewStyle
-}
-
-export type ExclusivityBannerProps = Omit<ExclusivityModuleProps, 'offerId' | 'url'>
 
 const UnmemoizedExclusivityModule = ({ offerId, url, ...props }: ExclusivityModuleProps) => {
   const ExclusivityComponent = ({ style }: { style?: ViewStyle }) => {
