+++ src/features/home/components/modules/marketing/MarketingBlock.tsx
@@ -3,17 +3,7 @@ import { useTheme } from 'styled-components/native'
 
 import { MarketingBlockContent } from 'features/home/components/modules/marketing/MarketingBlockContent'
 import { MarketingBlockContentDesktop } from 'features/home/components/modules/marketing/MarketingBlockContentDesktop'
-import { InternalNavigationProps } from 'ui/components/touchableLink/types'
-
-export type MarketingBlockProps = {
-  navigateTo: InternalNavigationProps['navigateTo']
-  onBeforeNavigate?: () => void
-  accessibilityLabel?: string
-  backgroundImageUrl?: string
-  AttachedCardComponent: React.ReactNode
-  withGradient?: boolean
-  gradientHeight?: string
-}
+import { MarketingBlockProps } from 'features/home/types'
 
 const GRADIENT_HEIGHT = '100%'
