+++ src/features/home/components/modules/marketing/MarketingBlockContentDesktop.tsx
@@ -3,7 +3,7 @@ import { View } from 'react-native'
 import styled from 'styled-components/native'
 
 import { BlackGradient } from 'features/home/components/BlackGradient'
-import { MarketingBlockProps } from 'features/home/components/modules/marketing/MarketingBlock'
+import { MarketingBlockProps } from 'features/home/types'
 import { BlurAmount } from 'ui/components/BlurryWrapper/BlurAmount'
 import { BlurryWrapper } from 'ui/components/BlurryWrapper/BlurryWrapper'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
