+++ src/features/offer/components/MovieCalendar/MovieCalendar.tsx
@@ -6,7 +6,10 @@ import styled, { useTheme } from 'styled-components/native'
 
 import { MovieCalendarBottomBar } from 'features/offer/components/MovieCalendar/components/MovieCalendarBottomBar'
 import { MovieCalendarDay } from 'features/offer/components/MovieCalendar/components/MovieCalendarDay'
-import { handleMovieCalendarScroll } from 'features/offer/components/MoviesScreeningCalendar/utils'
+import {
+  handleMovieCalendarScroll,
+  MOVIE_CALENDAR_PADDING,
+} from 'features/offer/components/MoviesScreeningCalendar/utils'
 import { useHorizontalFlatListScroll } from 'ui/hooks/useHorizontalFlatListScroll'
 import { PlaylistArrowButton } from 'ui/Playlist/PlaylistArrowButton'
 import { getSpacing } from 'ui/theme'
@@ -22,8 +25,6 @@ type Props = {
   itemWidth?: number
   onItemLayout?: (event: LayoutChangeEvent) => void
 }
-
-export const MOVIE_CALENDAR_PADDING = getSpacing(6)
 
 export const MovieCalendar: React.FC<Props> = ({
   dates,
