+++ src/features/profile/components/VerticalDots/VerticalDots.tsx
@@ -4,6 +4,8 @@ import styled from 'styled-components/native'
 // Disable ESLint because I need this colors enum.
 // eslint-disable-next-line no-restricted-imports
 
+import { DotSize, VerticalDotsProps } from 'features/profile/types'
+
 import { AutomaticVerticalDots } from './AutomaticVerticalDots'
 
 /**
@@ -11,49 +13,6 @@ import { AutomaticVerticalDots } from './AutomaticVerticalDots'
  * or an object with width and height that allows you to get
  * nice shapes.
  */
-export type DotSize = number | { width: number; height: number }
-
-export interface VerticalDotsProps {
-  /**
-   * Only used to compute dot size when `dotSize` is not passed.
-   */
-  parentWidth: number
-  /**
-   * This is the **VerticalDots** parent height.
-   *
-   * Used to compute dot count.
-   */
-  parentHeight: number
-  /**
-   * Manually set dot size.
-   *
-   * If not given, it will automatically set dot size based on parentWidth value.
-   *
-   * @default parentWidth
-   */
-  dotSize?: DotSize
-  /**
-   * Specifies if it should end with a dot.
-   *
-   * - When you only have one dotted line it should be set to `true`.
-   * - When you have multiple dotted lines following each other, it should be set to `false` so
-   * it seems linear.
-   */
-  endsWithDot?: boolean
-  /**
-   * If you want a custom size for the first dot.
-   *
-   * @default dotSize
-   */
-  firstDotSize?: DotSize
-  /**
-   * If you want a custom size for the last dot.
-   *
-   * @default dotSize
-   */
-  lastDotSize?: DotSize
-  testID?: string
-}
 
 type GetDotCountOptions = {
   availableHeight: number
