+++ src/features/profile/pages/NotificationSettings/NotificationsSettings.tsx
@@ -14,6 +14,7 @@ import {
   hasUserChangedSubscriptions,
 } from 'features/profile/pages/NotificationSettings/helpers/hasUserChangedParameters'
 import { usePushPermission } from 'features/profile/pages/NotificationSettings/usePushPermission'
+import { NotificationsSettingsState } from 'features/profile/types'
 import { mapSubscriptionThemeToName } from 'features/subscription/helpers/mapSubscriptionThemeToName'
 import {
   SubscriptionTheme,
@@ -32,12 +33,6 @@ import { SecondaryPageWithBlurHeader } from 'ui/pages/SecondaryPageWithBlurHeade
 import { Info } from 'ui/svg/icons/Info'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
 import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
-
-export type NotificationsSettingsState = {
-  allowEmails?: boolean
-  allowPush?: boolean
-  themePreferences: SubscriptionTheme[]
-}
 
 export const NotificationsSettings = () => {
   const { isLoggedIn, user } = useAuthContext()
